chore: moved table ordering code into a seperate file
