+++ apps/web/src/app/(dashboard)/(management)/locatie/[location]/cohorten/_components/table.tsx
@@ -1,10 +1,4 @@
 "use client";
-import { closestCenter, DndContext, DragEndEvent } from "@dnd-kit/core";
-import { restrictToHorizontalAxis } from "@dnd-kit/modifiers";
-import {
-  horizontalListSortingStrategy,
-  SortableContext,
-} from "@dnd-kit/sortable";
 import {
   createColumnHelper,
   flexRender,
@@ -14,23 +8,25 @@ import {
 } from "@tanstack/react-table";
 import clsx from "clsx";
 import { useParams } from "next/navigation";
-import { useState } from "react";
 import {
   Table,
   TableBody,
   TableCell,
-  TableDisplay,
   TableHead,
   TableHeader,
   TableRow,
-  updateColumnOrder,
 } from "~/app/(dashboard)/_components/table";
 import {
   TableFooter,
   TablePagination,
   TableRowSelection,
 } from "~/app/(dashboard)/_components/table-footer";
-import { useCustomSensors } from "~/app/(dashboard)/_hooks/custom-sensors";
+import {
+  TableColumnOrderingContext,
+  TableDisplay,
+  TableOrderingContext,
+} from "~/app/(dashboard)/_components/table-ordering";
+import { useColumnOrdering } from "~/app/(dashboard)/_hooks/use-column-ordering";
 import dayjs from "~/lib/dayjs";
 import type { listCohortsForLocation } from "~/lib/nwd";
 
@@ -74,84 +70,24 @@ export default function CohortsTable({
   cohorts: Awaited<ReturnType<typeof listCohortsForLocation>>;
   totalItems: number;
 }) {
-  // For column ordering
-  // NOTE: We need to replace `.` with `_` because `.` is not a valid key in React Table.
-  const [columnOrder, setColumnOrder] = useState<string[]>(() =>
-    columns
-      .map((column) => column.accessorKey)
-      .map((column) => column.replace(/\./g, "_")),
-  );
-
-  // For column ordering
-  const [columnVisibility, setColumnVisibility] = useState<
-    Record<string, boolean>
-  >(
-    columns.reduce(
-      (
-        acc: Record<string, boolean>,
-        column: {
-          id?: string;
-          isDefaultVisible?: boolean;
-        },
-      ) => {
-        if (column.id !== undefined && column.isDefaultVisible !== undefined) {
-          acc[column.id] = column.isDefaultVisible;
-        }
-        return acc;
-      },
-      {},
-    ),
-  );
+  const options = useColumnOrdering({ columns });
 
   const table = useReactTable({
     data: cohorts,
     columns,
-    state: {
-      columnOrder, // For column ordering
-      columnVisibility, // For column ordering
-    },
     getRowId: (row) => row.id,
     getCoreRowModel: getCoreRowModel(),
     getSortedRowModel: getSortedRowModel(),
-    onColumnOrderChange: setColumnOrder, // For column ordering
-    onColumnVisibilityChange: setColumnVisibility, // For column ordering
+    ...options,
   });
-
-  const sensors = useCustomSensors(); // For column ordering
-
-  // For column ordering
-  function handleDragEnd(event: DragEndEvent) {
-    const { active, over } = event;
-
-    if (active && over && active.id !== over.id) {
-      const newColumnOrder = updateColumnOrder(
-        columnOrder,
-        active.id as string,
-        over.id as string,
-      );
-
-      setColumnOrder(newColumnOrder);
-    }
-  }
 
   const params = useParams();
 
   return (
     <>
-      {/* For column ordering */}
-      <DndContext
-        collisionDetection={closestCenter}
-        modifiers={[restrictToHorizontalAxis]}
-        onDragEnd={(event) => handleDragEnd(event)}
-        sensors={sensors}
-      >
-        {/* For column ordering */}
+      <TableOrderingContext options={options}>
         <div className="mt-8 flex justify-end items-center gap-1">
-          <TableDisplay
-            table={table}
-            columnOrder={columnOrder}
-            setColumnOrder={setColumnOrder}
-          />
+          <TableDisplay table={table} />
         </div>
         <Table
           className="mt-1 [--gutter:theme(spacing.6)] lg:[--gutter:theme(spacing.10)]"
@@ -160,11 +96,7 @@ export default function CohortsTable({
           <TableHead>
             {table.getHeaderGroups().map((headerGroup) => (
               <TableRow key={headerGroup.id}>
-                {/* For column ordering */}
-                <SortableContext
-                  items={columnOrder}
-                  strategy={horizontalListSortingStrategy}
-                >
+                <TableColumnOrderingContext>
                   {headerGroup.headers.map((header) => (
                     <TableHeader
                       key={header.id}
@@ -177,7 +109,7 @@ export default function CohortsTable({
                       )}
                     </TableHeader>
                   ))}
-                </SortableContext>
+                </TableColumnOrderingContext>
               </TableRow>
             ))}
           </TableHead>
@@ -199,11 +131,7 @@ export default function CohortsTable({
                 key={row.id}
                 href={`/locatie/${params.location as string}/cohorten/${row.original.handle}`}
               >
-                {/* For column ordering */}
-                <SortableContext
-                  items={columnOrder}
-                  strategy={horizontalListSortingStrategy}
-                >
+                <TableColumnOrderingContext>
                   {row.getVisibleCells().map((cell, index) => (
                     <TableCell
                       key={cell.id}
@@ -221,12 +149,12 @@ export default function CohortsTable({
                       )}
                     </TableCell>
                   ))}
-                </SortableContext>
+                </TableColumnOrderingContext>
               </TableRow>
             ))}
           </TableBody>
         </Table>
-      </DndContext>
+      </TableOrderingContext>
 
       <TableFooter>
         <TableRowSelection table={table} totalItems={totalItems} />
