+++ apps/web/src/app/(dashboard)/_components/table.tsx
@@ -1,40 +1,11 @@
 "use client";
 
-import { closestCenter, DndContext, DragEndEvent } from "@dnd-kit/core";
-import { restrictToVerticalAxis } from "@dnd-kit/modifiers";
-import {
-  arrayMove,
-  SortableContext,
-  useSortable,
-  verticalListSortingStrategy,
-} from "@dnd-kit/sortable";
+import { useSortable } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
-import { EyeIcon, EyeSlashIcon } from "@heroicons/react/24/outline";
-import {
-  Cell,
-  Column,
-  Header,
-  Table as TableType,
-} from "@tanstack/react-table";
+import { Cell, Header } from "@tanstack/react-table";
 import { clsx } from "clsx";
 import type React from "react";
-import {
-  ComponentProps,
-  createContext,
-  CSSProperties,
-  useContext,
-  useState,
-} from "react";
-import DragIcon from "~/app/_components/drag-icon";
-import { useCustomSensors } from "../_hooks/custom-sensors";
-import { Checkbox } from "./checkbox";
-import {
-  Dropdown,
-  DropdownButton,
-  DropdownItem,
-  DropdownMenu,
-  DropdownSection,
-} from "./dropdown";
+import { createContext, CSSProperties, useContext, useState } from "react";
 import { Link } from "./link";
 
 const TableContext = createContext<{
@@ -184,7 +155,6 @@ export function TableHeader<TData, _>({
     <th
       {...props}
       colSpan={header.colSpan}
-      ref={setNodeRef} // For column ordering
       className={clsx(
         className,
         "border-b border-b-zinc-950/10 px-4 py-2 font-medium first:pl-[var(--gutter,theme(spacing.2))] last:pr-[var(--gutter,theme(spacing.2))] dark:border-b-white/10",
@@ -192,9 +162,11 @@ export function TableHeader<TData, _>({
           "border-l border-l-zinc-950/5 first:border-l-0 dark:border-l-white/5",
         !bleed && "sm:first:pl-1 sm:last:pr-1",
       )}
-      style={style} // For column ordering
-      {...attributes} // For column ordering
-      {...listeners} // For column ordering
+      // Column Ordering
+      ref={setNodeRef}
+      style={style}
+      {...attributes}
+      {...listeners}
     />
   );
 }
@@ -241,7 +213,8 @@ export function TableCell<TData, _>({
         dense ? "py-2.5" : "py-4",
         !bleed && "sm:first:pl-1 sm:last:pr-1",
       )}
-      style={style} // For column ordering
+      // Column ordering
+      style={style}
     >
       {href && (
         <Link
@@ -255,171 +228,5 @@ export function TableCell<TData, _>({
       )}
       {children}
     </td>
-  );
-}
-
-// For column ordering
-export function updateColumnOrder(
-  columnOrder: string[],
-  activeId: string,
-  overId: string,
-): string[] {
-  const oldIndex = columnOrder.indexOf(activeId);
-  const newIndex = columnOrder.indexOf(overId);
-  return arrayMove(columnOrder, oldIndex, newIndex);
-}
-
-// For column ordering
-export function TableDisplay<TData>({
-  table,
-  columnOrder,
-  setColumnOrder,
-  anchor = "bottom end",
-}: {
-  table: TableType<TData>;
-  columnOrder: string[];
-  setColumnOrder: (columnOrder: string[]) => void;
-  anchor?: ComponentProps<typeof DropdownMenu>["anchor"];
-}) {
-  const sensors = useCustomSensors();
-  const isLastColumnVisible = table.getVisibleLeafColumns().length === 1;
-
-  function handleDragEnd(event: DragEndEvent) {
-    const { active, over } = event;
-
-    if (active && over && active.id !== over.id) {
-      const newColumnOrder = updateColumnOrder(
-        columnOrder,
-        active.id as string,
-        over.id as string,
-      );
-
-      setColumnOrder(newColumnOrder);
-    }
-  }
-
-  return (
-    <Dropdown>
-      <DropdownButton
-        outline
-        className={clsx(
-          !table.getIsAllColumnsVisible() && "border-branding-dark/10",
-        )}
-      >
-        <span
-          className={clsx(
-            "hidden sm:inline",
-            !table.getIsAllColumnsVisible() && "text-branding-dark",
-          )}
-        >
-          Weergave
-        </span>
-        {!table.getIsAllColumnsVisible() ? (
-          <EyeSlashIcon className="h-4 w-4 [--btn-icon:theme(colors.branding.dark)]" />
-        ) : (
-          <EyeIcon className="h-4 w-4" />
-        )}
-      </DropdownButton>
-      <DropdownMenu anchor={anchor} className="min-w-64">
-        <DropdownSection>
-          <DndContext
-            collisionDetection={closestCenter}
-            modifiers={[restrictToVerticalAxis]}
-            onDragEnd={(event) => handleDragEnd(event)}
-            sensors={sensors}
-          >
-            {table.getAllLeafColumns().map((column) => {
-              return (
-                <SortableContext
-                  key={column.id}
-                  items={columnOrder}
-                  strategy={verticalListSortingStrategy}
-                >
-                  <SortableItem
-                    column={column as Column<unknown, unknown>}
-                    isDisabled={column.getIsVisible() && isLastColumnVisible}
-                  />
-                </SortableContext>
-              );
-            })}
-          </DndContext>
-        </DropdownSection>
-        <DropdownSection>
-          <DropdownItem
-            className={clsx(
-              "!grid-cols-[1fr_1.5rem_0.5rem_auto]",
-              table.getIsAllColumnsVisible()
-                ? "cursor-not-allowed opacity-60"
-                : "cursor-pointer opacity-100",
-            )}
-            onClick={() => {
-              if (!table.getIsAllColumnsVisible())
-                table.toggleAllColumnsVisible();
-            }}
-          >
-            <span className="font-medium">Toon alles</span>
-            <Checkbox
-              disabled={table.getIsAllColumnsVisible()}
-              checked={table.getIsAllColumnsVisible()}
-              onChange={() => {
-                if (!table.getIsAllColumnsVisible())
-                  table.toggleAllColumnsVisible();
-              }}
-            />
-          </DropdownItem>
-        </DropdownSection>
-      </DropdownMenu>
-    </Dropdown>
-  );
-}
-
-function SortableItem({
-  column,
-  isDisabled,
-}: {
-  column: Column<unknown, unknown>;
-  isDisabled: boolean;
-}) {
-  const {
-    isDragging,
-    attributes,
-    listeners,
-    setNodeRef,
-    transform,
-    transition,
-  } = useSortable({ id: column?.id });
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition,
-  };
-
-  return (
-    <DropdownItem
-      ref={setNodeRef}
-      style={style}
-      {...attributes}
-      {...listeners}
-      type="button"
-      onClick={() => !isDisabled && column.toggleVisibility()}
-      className={clsx(
-        isDisabled ? "opacity-50" : "opacity-100",
-        isDragging && "z-10", // This is needed to prevent other items accidentally from being dragged
-      )}
-      as={"div"}
-    >
-      <DragIcon isDragging={isDragging} />
-      <span>
-        {typeof column.columnDef.header === "function"
-          ? (column.columnDef.meta as { label?: string })?.label
-          : column.columnDef.header}
-      </span>
-
-      <Checkbox
-        disabled={isDisabled}
-        checked={column.getIsVisible()}
-        onChange={() => column.toggleVisibility()}
-      />
-    </DropdownItem>
   );
 }
